Split nsot/models.py to one model per file
This breaks up the models.py monolith so that each model (and its
associated other bits) each live in their own file. Hopefully this makes
things easier to read, understand, and PRs easier to review.

One downside is that we lose git history on each of these files, but
this commit message hopefully serves as a good redirectoin to find the
original file, and I think the benefits outweigh that negative.

Fixes #248
